Change 'value' to uint32_t in pm4/sdma stream decoding callbacks
The "add_field()" callback was using a uint64_t for the "value"
field which is 32-bit by definition.  Switch to uint32_t for
clarity.

+++ src/lib/umr_sdma_decode_opcodes.c
@@ -337,7 +337,7 @@ static void start_opcode(struct umr_sdma_stream_decode_ui *ui, uint64_t ib_addr,
 	printf("Opcode 0x%lx [%s] at %lu@[0x%llx + 0x%llx] (%lu words)\n", (unsigned long)opcode, opcode_name, (unsigned long)ib_vmid, (unsigned long long)data->off[data->i - 1], (unsigned long long)ib_addr - data->off[data->i - 1], (unsigned long)nwords);
 	fflush(stdout);
 }
-static void add_field(struct umr_sdma_stream_decode_ui *ui, uint64_t ib_addr, uint32_t ib_vmid, const char *field_name, uint64_t value, char *str, int ideal_radix)
+static void add_field(struct umr_sdma_stream_decode_ui *ui, uint64_t ib_addr, uint32_t ib_vmid, const char *field_name, uint32_t value, char *str, int ideal_radix)
 {
 	struct demo_ui_data *data = ui->data;
 	printf("\t[%lu@0x%llx + 0x%llx] -- %s == ", (unsigned long)ib_vmid, (unsigned long long)data->off[data->i - 1], (unsigned long long)ib_addr - data->off[data->i - 1], field_name);
